+++ cli/integration-test/sample-test/src/sample_test/diagnostics/depend/a.clj
@@ -1,0 +1,3 @@
+(ns sample-test.diagnostics.depend.a)
+
+(def foo 1)