+++ cli/integration-test/sample-test/src/sample_test/diagnostics/depend/b.clj
@@ -1,0 +1,4 @@
+(ns sample-test.diagnostics.depend.b
+  (:require [sample-test.diagnostics.depend.a :as depend.a]))
+
+depend.a/foo