+++ lib/src/clojure_lsp/feature/file_management.clj
@@ -1,5 +1,6 @@
 (ns clojure-lsp.feature.file-management
   (:require
+   [clojure-lsp.clj-depend :as lsp.depend]
    [clojure-lsp.clojure-producer :as clojure-producer]
    [clojure-lsp.crawler :as crawler]
    [clojure-lsp.db :as db]
@@ -36,9 +37,15 @@
   (swap! db* #(assoc-in % [:documents uri] {:v 0 :text text :saved-on-disk false})))
 
 (defn did-open [uri text db* allow-create-ns]
-  (when-let [kondo-result (lsp.kondo/run-kondo-on-text! text uri db*)]
-    (load-document! uri text db*)
-    (swap! db* #(lsp.kondo/db-with-results % kondo-result))
+  (load-document! uri text db*)
+  (let [kondo-result* (future (lsp.kondo/run-kondo-on-text! text uri db*))
+        depend-result* (future (lsp.depend/analyze-filename! (shared/uri->filename uri) @db*))
+        kondo-result @kondo-result*
+        depend-result @depend-result*]
+    (swap! db* (fn [state-db]
+                 (-> state-db
+                     (lsp.kondo/db-with-results kondo-result)
+                     (lsp.depend/db-with-results depend-result))))
     (f.diagnostic/async-publish-diagnostics! uri @db*))
   (when allow-create-ns
     (when-let [create-ns-edits (create-ns-changes uri text @db*)]
@@ -211,16 +218,24 @@
       new-text)))
 
 (defn analyze-changes [{:keys [uri text version]} {:keys [producer db*] :as components}]
-  (loop [state-db @db*]
-    (when (>= version (get-in state-db [:documents uri :v] -1))
-      (when-let [kondo-result (shared/logging-time
+  (let [filename (shared/uri->filename uri)]
+    (loop [state-db @db*]
+      (when (>= version (get-in state-db [:documents uri :v] -1))
+        (let [kondo-result* (future
+                              (shared/logging-time
                                 (str "changes analyzed by clj-kondo took %s")
-                                (lsp.kondo/run-kondo-on-text! text uri db*))]
-        (let [filename (shared/uri->filename uri)
+                                (lsp.kondo/run-kondo-on-text! text uri db*)))
+              depend-result* (future
+                               (shared/logging-time
+                                 (str "changes analyzed by clj-depend took %s")
+                                 (lsp.depend/analyze-filename! filename state-db)))
+              kondo-result @kondo-result*
+              depend-result @depend-result*
               old-db @db*]
           (if (compare-and-set! db* state-db
                                 (-> state-db
                                     (lsp.kondo/db-with-results kondo-result)
+                                    (lsp.depend/db-with-results depend-result)
                                     (update :processing-changes disj uri)))
             (let [db @db*]
               (f.diagnostic/sync-publish-diagnostics! uri db)